ingest_function/requirements.txt
requests==2.31.0
google-cloud-bigquery==3.11.4